Rename Renderer to AudioPlayer
