+++ src/main/java/sh/ball/audio/FrameProducer.java
@@ -2,13 +2,13 @@ package sh.ball.audio;
 
 public class FrameProducer<S, T> implements Runnable {
 
-  private final Renderer<S, T> renderer;
+  private final AudioPlayer<S, T> audioPlayer;
   private final FrameSet<S> frames;
 
   private boolean running;
 
-  public FrameProducer(Renderer<S, T> renderer, FrameSet<S> frames) {
-    this.renderer = renderer;
+  public FrameProducer(AudioPlayer<S, T> audioPlayer, FrameSet<S> frames) {
+    this.audioPlayer = audioPlayer;
     this.frames = frames;
   }
 
@@ -16,7 +16,7 @@ public class FrameProducer<S, T> implements Runnable {
   public void run() {
     running = true;
     while (running) {
-      renderer.addFrame(frames.next());
+      audioPlayer.addFrame(frames.next());
     }
   }
